+++ packages/events/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @opengeni/events
 
+## 0.3.56
+
+### Patch Changes
+
+- Updated dependencies [dd71248]
+- Updated dependencies [03ed7eb]
+  - @opengeni/contracts@0.29.0
+  - @opengeni/db@0.20.0
+
 ## 0.3.55
 
 ### Patch Changes