+++ packages/github/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @opengeni/github
 
+## 0.4.14
+
+### Patch Changes
+
+- Updated dependencies [dd71248]
+  - @opengeni/contracts@0.29.0
+  - @opengeni/config@0.9.2
+
 ## 0.4.13
 
 ### Patch Changes